backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
blinker==1.9.0
    # via flask
blis==1.3.0
    # via thinc
cachetools==6.2.0
    # via google-auth
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2025.10.5
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via
    #   flask
    #   typer
cloudpathlib==0.23.0
    # via weasel
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   wasabi
confection==0.1.5
    # via
    #   thinc
    #   weasel
cymem==2.0.11
    # via
    #   preshed
    #   spacy
    #   thinc
dnspython==2.8.0
    # via pymongo
flask==3.1.2
    # via
    #   -r requirements.in
    #   flask-cors
flask-cors==6.0.1
    # via -r requirements.in
google-api-core==2.26.0
    # via google-api-python-client
google-api-python-client==2.184.0
    # via -r requirements.in
google-auth==2.41.1
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-auth-oauthlib==1.2.2
    # via -r requirements.in
googleapis-common-protos==1.70.0
    # via google-api-core
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   spacy
langcodes==3.5.0
    # via spacy
language-data==1.3.0
    # via langcodes
marisa-trie==1.3.1
    # via language-data
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
murmurhash==1.0.13
    # via
    #   preshed
    #   spacy
    #   thinc
numpy==2.3.4
    # via
    #   blis
    #   spacy
    #   thinc
oauthlib==3.3.1
    # via requests-oauthlib
packaging==25.0
    # via
    #   spacy
    #   thinc
    #   weasel
preshed==3.0.10
    # via
    #   spacy
    #   thinc
proto-plus==1.26.1
    # via google-api-core
protobuf==6.32.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.3
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.41.4
    # via pydantic
pygments==2.19.2
    # via rich
pymongo==4.15.3
    # via -r requirements.in
pyparsing==3.2.5
    # via httplib2
python-dotenv==1.1.1
    # via -r requirements.in
requests==2.32.5
    # via
    #   -r requirements.in
    #   google-api-core
    #   requests-oauthlib
    #   spacy
    #   weasel
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rich==14.2.0
    # via typer
rsa==4.9.1
    # via google-auth
shellingham==1.5.4
    # via typer
smart-open==7.4.1
    # via weasel
spacy==3.8.7
    # via -r requirements.in
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
srsly==2.5.1
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
thinc==8.3.6
    # via spacy
tqdm==4.67.1
    # via spacy
typer==0.20.0
    # via
    #   spacy
    #   weasel
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
weasel==0.4.1
    # via spacy
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
wrapt==2.0.0
    # via smart-open

# The following packages are considered to be unsafe in a requirements file:
# setuptools
